moogle/moogle/requirements/local.txt
# Extend base requirements
-r base.txt

django-debug-toolbar
# which requires:
# sqlparse
werkzeug
bpython
# which requires:
# pygments
django-extensions
# which requires:
# six